Refactor around some functions and add comments
