+++ docker/test.dockerfile
@@ -1,8 +1,6 @@
 ARG BASE_IMAGE=${BASE_IMAGE:-3.7-slim-stretch}
 FROM python:${BASE_IMAGE}
 
-MAINTAINER Cung Tran "minishcung@gmail.com"
-
 # Install BLAS for numpy and OpenMP for faiss
 RUN apt update && \
         apt install -y libopenblas-dev libomp-dev make && \
