+++ needlestack/servicers/searcher.py
@@ -11,7 +11,7 @@ from needlestack.collections.collection import Collection
 from needlestack.collections.shard import Shard
 from needlestack.cluster_managers import ClusterManager
 from needlestack.servicers.settings import BaseConfig
-from needlestack.servicers.decorators import unhandled_exception_rpc
+from needlestack.utilities.rpc import unhandled_exception_rpc
 
 
 logger = logging.getLogger("needlestack")
